feat : Create Item Equip  / fix : Inventory Error / feat : font change
폰트 변경
인벤토리 스택 오류 처리
아이템 장착 로직 추가
  ㄴ 구글시트 CSV 에 들수있는 아이템 변수를 추가 및 활용하여 작업 완료

+++ Assets/Resources/Scripts/Runtime/Game_Scripts/Interact_Scripts/Container_Scripts/Interact_Container.cs
@@ -118,7 +118,7 @@ public class Interact_Container : MonoBehaviour, IInteract
         ItemDataSO item = slot.GetItem; // 슬롯의 아이템
         int amount = slot.Count;     // 추가할 슬롯의 아이템 갯수
         bool isStack = item.IsStackable; // 아이템 스택유무
-        float dur = slot.Dur;
+        int dur = slot.Dur;
 
         if (isStack)
         {
